Working Firmware with the wps button on wifi board, almost working weight_master

+++ weightlogger_master/scale_i2c.cpp
@@ -7,7 +7,7 @@ ElcanScale::ElcanScale() {
   _i2c_address = 0;
 }
 
-bool ElcanScale::begin(int addr) {
+bool ElcanScale::begin(byte addr) {
   char tmp[8];
   char pos = 0;
   memset(tmp, 0, sizeof(tmp));
@@ -15,7 +15,7 @@ bool ElcanScale::begin(int addr) {
   Wire.beginTransmission(_i2c_address);
   Wire.write("I"); //send INIT
   Wire.endTransmission();
-  Wire.requestFrom(_i2c_address, 6);
+  Wire.requestFrom(_i2c_address, 6u);
   while (Wire.available()) {
     char b = Wire.read();
     if (b != 0xFF) {
