+++ weightlogger_master/scale_i2c.h
@@ -6,11 +6,11 @@ class ElcanScale {
   private:
     bool error;
     byte _error;
-    byte _i2c_address;
+    uint8_t _i2c_address;
 
   public:
     ElcanScale();
-    bool begin(int i2c_address);
+    bool begin(uint8_t i2c_address);
     uint16_t get_weight();
     bool is_error();
 };
